organises into component level folders
extract counter component to own folder
